+++ 1.21.11/src/main/java/com/bleudev/nine_lifes/mixin/client/EnvironmentAttributeProbeMixin.java
@@ -1,6 +1,6 @@
 package com.bleudev.nine_lifes.mixin.client;
 
-import com.bleudev.nine_lifes.client.FogProperitesKt;
+import com.bleudev.nine_lifes.api.event.client.ClientEnvironmentSetupEvents;
 import net.minecraft.world.attribute.EnvironmentAttribute;
 import net.minecraft.world.attribute.EnvironmentAttributeProbe;
 import net.minecraft.world.attribute.EnvironmentAttributes;
@@ -10,12 +10,14 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+@SuppressWarnings("unchecked")
 @Mixin(EnvironmentAttributeProbe.class)
 public class EnvironmentAttributeProbeMixin {
     @Inject(method = "getValue", at = @At("RETURN"), cancellable = true)
     private <Value> void modifySkyColor(EnvironmentAttribute<@NotNull Value> environmentAttribute, float f, CallbackInfoReturnable<Value> cir) {
         if (environmentAttribute.equals(EnvironmentAttributes.SKY_COLOR)) {
-            cir.setReturnValue((Value) FogProperitesKt.getSkyColor().invoke((Integer) cir.getReturnValue()));
+            int original = (Integer) cir.getReturnValue();
+            cir.setReturnValue((Value) ClientEnvironmentSetupEvents.SKY_COLOR.invoker().invoke(original, original));
         }
     }
 }