+++ 1.21.11/src/main/kotlin/com/bleudev/nine_lifes/api/event/client/ClientRespawnEvents.kt
@@ -1,13 +1,16 @@
 package com.bleudev.nine_lifes.api.event.client
 
+import net.fabricmc.api.EnvType
+import net.fabricmc.api.Environment
 import net.fabricmc.fabric.api.event.Event
 import net.fabricmc.fabric.api.event.EventFactory
 import net.minecraft.client.Minecraft
 
-private typealias ClientRespawn = (Minecraft) -> Unit
-
+@Environment(EnvType.CLIENT)
 object ClientRespawnEvents {
     val RESPAWN: Event<ClientRespawn> = EventFactory.createArrayBacked(ClientRespawn::class.java) {
         { mc -> for (i in it) i(mc) }
     }
+
+    typealias ClientRespawn = (Minecraft) -> Unit
 }