+++ 1.21.11/src/main/kotlin/com/bleudev/nine_lifes/client/render/NineLifesPostRenderer.kt
@@ -1,8 +1,8 @@
 package com.bleudev.nine_lifes.client.render
 
-import com.bleudev.nine_lifes.client.api.render.DynamicUniformsRegistryImpl
+import com.bleudev.nine_lifes.api.render.client.DynamicUniformsRegistryImpl
+import com.bleudev.nine_lifes.api.render.client.PostEffectRegistryImpl
 import com.bleudev.nine_lifes.client.util.applyPostEffect
-import com.bleudev.nine_lifes.util.createIdentifier
 import com.mojang.blaze3d.resource.CrossFrameResourcePool
 import net.minecraft.client.Minecraft
 import net.minecraft.resources.Identifier
@@ -14,11 +14,8 @@ class NineLifesPostRenderer(val resourcePool: CrossFrameResourcePool, val minecr
 
     fun render() {
         DynamicUniformsRegistryImpl.updateBuffers()
-        renderCustomPostEffect("redmaj")
-        renderCustomPostEffect("anaglyph")
-        renderCustomPostEffect("cblur")
+        PostEffectRegistryImpl.execute(::renderPostEffect)
     }
 
-    private fun renderCustomPostEffect(name: String) = renderPostEffect(createIdentifier(name))
     private fun renderPostEffect(postEffectId: Identifier) = minecraftSupplier().applyPostEffect(postEffectId, resourcePool)
 }